Move stake operations to the state crate
